Moved content serializers inside Content namespace
