+++ app/controllers/content/site_dives_controller.rb
@@ -4,7 +4,9 @@ module Content
 
     def show
       @site_dive = Content::SiteDive.find(params[:id])
-      render json: @site_dive, root: 'data', serializer: SiteDiveSerializer
+      render json: @site_dive,
+             root: 'data',
+             serializer: Content::SiteDiveSerializer
     end
   end
 end
