+++ app/controllers/content/testimonials_controller.rb
@@ -4,7 +4,9 @@ module Content
 
     def index
       testimonials = Content::Testimonial.order(created_at: 'DESC')
-      render json: testimonials, root: 'data', each_serializer: TestimonialSerializer
+      render json: testimonials,
+             root: 'data',
+             each_serializer: Content::TestimonialSerializer
     end
   end
 end
