+++ app/serializers/content/site_dive_serializer.rb
@@ -1,0 +1,5 @@
+module Content
+  class SiteDiveSerializer < ActiveModel::Serializer
+    attributes :title, :description
+  end
+end
