--- app/serializers/site_dive_serializer.rb
@@ -1,3 +1,0 @@
-class SiteDiveSerializer < ActiveModel::Serializer
-  attributes :title, :description
-end
